fix: restore missing Lombok imports + fix compilation errors
- Add missing 'import lombok.Data' and 'import lombok.NoArgsConstructor'
  to 22 Entity files across all modules (product, notification,
  payment, storage, profile, shopinfo, cart)
  These were silently removed by Spotless formatter
- Add missing DocumentDomainRepository import in elasticsearch module
- Fix int-to-Long type mismatch in DashboardServiceImpl

All 12 modules now compile successfully (Java 21).

+++ dashboard/src/main/java/com/fourman/dashboard/service/DashboardServiceImpl.java
@@ -135,7 +135,7 @@ public class DashboardServiceImpl implements DashboardService {
                         entry.getValue().stream()
                                 .mapToLong(OrderDTO::getTotalPrice)
                                 .sum(),
-                        entry.getValue().size()))
+                        (long) entry.getValue().size()))
                 .collect(Collectors.toList());
     }
